requirements.txt
Django==5.0.3
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-js-asset==2.2.0
fonttools==4.50.0
html5lib==1.1
lottie==0.7.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.21
pydyf==0.9.0
pyphen==0.14.0
python-decouple==3.8
python-slugify==8.0.4
text-unidecode==1.3
tinycss2==1.2.1
typing_extensions==4.10.0
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
